requirements.txt
# Core dependencies
google-generativeai
langchain-community
langchain-core
langchain-text-splitters
faiss-cpu
python-dotenv

# Benchmarking (optional — only needed for src.agent.benchmark)
pandas
openpyxl

# PDF extraction (optional — only needed for src.extraction)
llama-cloud
